Basic code for starting work on server and fixed build.gradle dependency typo
Added a number of files to get the basic server going, but still
requires filling in to populate the Request object and then
the Exchange object to manage the request. Will require handlers
to handle the requests as well. Added DataGenerator and StockDAO
placeholder files to be filled in as well.